+++ js/main.js
@@ -9,3 +9,4 @@ import './pictures.js';
 import './big-pictures.js';
 import './form-upload.js';
 import './api.js';
+import './slider.js';
